requirements.txt
Flask==2.3.6
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
